requirements.txt
paho_mqtt
numpy
mpu